Create CodeWars
This is a code from codewars.com. The goal was to write a function that takes an integer as input, and returns the number of bits that are equal to one in the binary representation of that number. The difficulty level was suppose to be easy but this was anything but that. I need more practice.